requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.7,<4.0
psycopg2>=2.8,<3.0
Pillow>=8.2,<9.0
pytest-django>=4.4.0,<5.0.0
